requirements.txt
fastapi
uvicorn
pydantic
requests
litellm
python-dotenv
numpy